+++ kiexport.py
@@ -4,8 +4,8 @@
 # KiExport
 # Tool to export manufacturing files from KiCad PCB projects.
 # Author: Vishnu Mohanan (@vishnumaiea, @vizmohanan)
-# Version: 0.0.10
-# Last Modified: +05:30 00:16:38 AM 31-08-2024, Saturday
+# Version: 0.0.11
+# Last Modified: +05:30 10:41:30 AM 01-09-2024, Sunday
 # GitHub: https://github.com/vishnumaiea/KiExport
 # License: MIT
 
@@ -410,6 +410,86 @@ def generatePcbPdf (output_dir, pcb_filename, to_overwrite = True):
 
 #=============================================================================================#
 
+def generateSchPdf (output_dir, sch_filename, to_overwrite = True):
+  # Common base command
+  sch_pdf_export_command = ["kicad-cli", "sch", "export", "pdf"]
+
+  if not check_file_exists (sch_filename):
+    print (f"generateSchPdf [ERROR]: {sch_filename} does not exist.")
+    return
+
+  file_name = extract_pcb_file_name (sch_filename)
+  project_name = extract_project_name (file_name)
+  info = extract_info_from_pcb (sch_filename)
+  
+  print (f"generateSchPdf [INFO]: Project name is {project_name} and revision is {info ['rev']}.")
+  
+  # Check if the ouptut directory exists, and create if not.
+  if not os.path.exists (output_dir):
+    print (f"generateSchPdf [INFO]: Output directory {output_dir} does not exist. Creating it now.")
+    os.makedirs (output_dir)
+
+  rev_directory = f"{output_dir}/R{info ['rev']}"
+
+  if not os.path.exists (rev_directory):
+    print (f"generateSchPdf [INFO]: Revision directory {rev_directory} does not exist. Creating it now.")
+    os.makedirs (rev_directory)
+  
+  not_completed = True
+  seq_number = 0
+  
+  while not_completed:
+    today_date = datetime.now()
+    formatted_date = today_date.strftime ("%d-%m-%Y")
+    filename_date = today_date.strftime ("%d%m%Y")
+    seq_number += 1
+    date_directory = f"{rev_directory}/[{seq_number}] {formatted_date}"
+    target_directory = f"{date_directory}/SCH"
+
+    if not os.path.exists (target_directory):
+      print (f"generateSchPdf [INFO]: Target directory {target_directory} does not exist. Creating it now.")
+      os.makedirs (target_directory)
+      not_completed = False
+    else:
+      if to_overwrite:
+        print (f"generateSchPdf [INFO]: Target directory {target_directory} already exists. Any files will be overwritten.")
+        not_completed = False
+      else:
+        print (f"generateSchPdf [INFO]: Target directory {target_directory} already exists. Creating another one.")
+        not_completed = True
+
+  # # Check if the target directory ends with a slash, and add one if not
+  # if target_directory [-1] != '/':
+  #   target_directory += '/'
+  
+  seq_number = 1
+  not_completed = True
+  
+  while not_completed:
+    file_name = f"{target_directory}/{project_name}-R{info ['rev']}-{type}-{filename_date}-{seq_number}.pdf"
+
+    if os.path.exists (file_name):
+      seq_number += 1
+      not_completed = True
+    else:
+      full_command = sch_pdf_export_command + \
+                    ["--output", f"{target_directory}/{project_name}-R{info ['rev']}-SCH-{filename_date}-{seq_number}.pdf"] + \
+                    ["--theme", "User"] + \
+                    [sch_filename]
+      not_completed = False
+  
+  # Run the command
+  try:
+    subprocess.run (full_command, check = True)
+  
+  except subprocess.CalledProcessError as e:
+    print (f"generateSchPdf [ERROR]: Error occurred: {e}")
+    return
+
+  print ("generateSchPdf [OK]: Schematic PDF file exported successfully.")
+
+#=============================================================================================#
+
 def generate3D (output_dir, pcb_filename, type, to_overwrite = True):
   # Common base command
   if type == "STEP" or type == "step":
@@ -625,24 +705,34 @@ def parseArguments():
   subparsers = parser.add_subparsers (dest = "command", help = "Available commands.")
 
   # Subparser for the Gerber export command
+  # Example: python .\kiexport.py gerbers -od "Mitayi-Pico-D1/Export" -if "Mitayi-Pico-D1/Mitayi-Pico-RP2040.kicad_pcb"
   gerber_parser = subparsers.add_parser ("gerbers", help = "Export Gerber files.")
   gerber_parser.add_argument ("-if", "--input_filename", required = True, help = "Path to the .kicad_pcb file.")
   gerber_parser.add_argument ("-od", "--output_dir", required = True, help = "Directory to save the Gerber files to.")
 
   # Subparser for the Position file export command
+  # Example: python .\kiexport.py positions -od "Mitayi-Pico-D1/Export" -if "Mitayi-Pico-D1/Mitayi-Pico-RP2040.kicad_pcb"
   position_parser = subparsers.add_parser ("positions", help = "Export Position files.")
   position_parser.add_argument ("-if", "--input_filename", required = True, help = "Path to the .kicad_pcb file.")
   position_parser.add_argument ("-od", "--output_dir", required = True, help = "Directory to save the Position files to.")
 
   # Subparser for the PCB PDF export command
+  # Example: python .\kiexport.py pcb_pdf -od "Mitayi-Pico-D1/Export" -if "Mitayi-Pico-D1/Mitayi-Pico-RP2040.kicad_pcb"
   pcb_pdf_parser = subparsers.add_parser ("pcb_pdf", help = "Export PCB PDF files.")
   pcb_pdf_parser.add_argument ("-if", "--input_filename", required = True, help = "Path to the .kicad_pcb file.")
   pcb_pdf_parser.add_argument ("-od", "--output_dir", required = True, help = "Directory to save the PCB PDF files to.")
 
-  # Subparser for the STEP export command
+  # Subparser for the Schematic PDF export command
+  # Example: python .\kiexport.py sch_pdf -od "Mitayi-Pico-D1/Export" -if "Mitayi-Pico-D1/Mitayi-Pico-RP2040.kicad_sch"
+  sch_pdf_parser = subparsers.add_parser ("sch_pdf", help = "Export schematic PDF files.")
+  sch_pdf_parser.add_argument ("-if", "--input_filename", required = True, help = "Path to the .kicad_sch file.")
+  sch_pdf_parser.add_argument ("-od", "--output_dir", required = True, help = "Directory to save the schematic PDF files to.")
+
+  # Subparser for the 3D file export command
+  # Example: python .\kiexport.py ddd -t "VRML" -od "Mitayi-Pico-D1/Export" -if "Mitayi-Pico-D1/Mitayi-Pico-RP2040.kicad_pcb"
   ddd_parser = subparsers.add_parser ("ddd", help = "Export 3D files.")
   ddd_parser.add_argument ("-if", "--input_filename", required = True, help = "Path to the .kicad_pcb file.")
-  ddd_parser.add_argument ("-od", "--output_dir", required = True, help = "Directory to save the STEP files to.")
+  ddd_parser.add_argument ("-od", "--output_dir", required = True, help = "Directory to save the 3D files to.")
   ddd_parser.add_argument ("-t", "--type", required = True, help = "The type of file to generate. Can be STEP or VRML.")
 
   test_parser = subparsers.add_parser ("test", help = "Test.")
@@ -661,6 +751,10 @@ def parseArguments():
   elif args.command == "pcb_pdf":
     # Call the generatePCBPDF function with the parsed arguments
     generatePcbPdf (args.output_dir, args.input_filename)
+
+  elif args.command == "sch_pdf":
+    # Call the generatePCBPDF function with the parsed arguments
+    generateSchPdf (args.output_dir, args.input_filename)
   
   elif args.command == "ddd":
     # Call the generateStep function with the parsed arguments